basic: lookup(): tokens are uint8_t
NB: return value needs to remain signed because -1 indicates "not found".

+++ ttbasic/basic.cpp
@@ -603,8 +603,8 @@ BString getstr() {
 //[戻り値]
 //  該当なし   : -1
 //  見つかった : キーワードコード
-int16_t lookup(char* str) {
-  for (int i = 0; i < SIZE_KWTBL; ++i) {
+int lookup(char* str) {
+  for (uint8_t i = 0; i < SIZE_KWTBL; ++i) {
     if (kwtbl[i] && !strncasecmp_P(str, kwtbl[i], strlen_P(kwtbl[i])))
       return i;
   }
